+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 
 ## [2.76.0] 2022-04-03
 
-- New command **sfdx **hardis:org:user:activateinvalid** to activate invalid emails in sandbox
+- New command **sfdx**hardis:org:user:activateinvalid** to activate invalid emails in sandbox
 - Fix CI org authentication in case the default username is not the org that we want to configure
 - Bypass error with force:source:legacy:pull / push
 - hardis:work:save : Propose to manually commit files